requirements.txt
anndata
biopython
imageio
ipython
jupyterlab
matplotlib
metapredict
networkx
numpy
openpyxl
pandas
requests
scanpy
scipy
scikit-image
scikit-learn
seaborn
sknetwork
tifffile
